Replace the embedded SQLKit utilities with use of the sql-kit-extras package.
